requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
fonttools==4.33.3
google-api-core==2.7.2
google-api-python-client==2.46.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-core==2.3.0
google-cloud-dataproc==4.0.2
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
httplib2==0.20.4
idna==3.3
kaggle==1.5.12
kaleido==0.2.1
kiwisolver==1.4.2
matplotlib==3.5.1
numpy==1.22.3
opendatasets==0.1.22
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pip==22.0.4
plotly==5.7.0
proto-plus==1.20.3
protobuf==3.20.1
py4j==0.10.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
pyspark==3.2.1
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
rsa==4.8
scipy==1.8.0
seaborn==0.11.2
setuptools==61.3.1
six==1.16.0
tenacity==8.0.1
text-unidecode==1.3
tqdm==4.64.0
uritemplate==4.1.1
urllib3==1.26.9
wheel==0.37.1
